Restructure folders so gen_coeff is at top
* move gen_coeff.py to gen_coeff/ and diode/ to gen_coeff/diode.py
* only use hybridj solver (fsolve) for now, since lm is sucky w/o
derivatives
* turns out everything is sucky without derivatives
* change fjrsh from rsh + 1/didv to vd(1/rsh + didv) so units are in
amps just like other residuals, where vd = isc*rs therefore vd/rsh = ish
* add weights parameter, default is 1.0
